Rename to pages and scripts
