+++ frontend/src/pages/PagesAndScriptsSettings/Components/PageTable/PageTable.style.tsx
