+++ frontend/src/pages/PagesAndScriptsSettings/Components/PageTable/index.tsx
