+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/AddScript/AddScript.tsx
