+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/AddScript/index.tsx
