+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/ScriptModal/ScriptModal.style.ts
