+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/ScriptModal/index.tsx
