+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/ScriptRow/ScriptRow.tsx
