+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/ScriptRow/index.tsx
