+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/ScriptTable.style.tsx
