+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/ScriptTableHeader/ScriptTableHeader.tsx
