+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/ScriptTableHeader/index.tsx
