+++ frontend/src/pages/PagesAndScriptsSettings/Components/ScriptTable/index.tsx
