+++ frontend/src/pages/PagesAndScriptsSettings/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './PagesAndScriptsSettings.wrap';
