--- frontend/src/pages/ScriptsAndPagesSettings/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './ScriptsAndPagesSettings.wrap';
