+++ frontend/src/translations/en.json
@@ -55,7 +55,7 @@
     "name": "Project name",
     "wpt_key": "WebPageTest API Key",
     "project_audit_parameters": "Project environments",
-    "scripts_and_pages": "Project scripts and pages",
+    "pages_and_scripts": "Project pages and scripts",
     "scripts": "Project scripts",
     "script": "Script",
     "script_name": "Script name",
